speed up stage tests No-Issue
