+++ galaxy_ng/tests/integration/api/test_artifact_upload.py
@@ -9,6 +9,7 @@ from orionutils.generator import build_collection, randstr
 from pkg_resources import parse_version
 
 from galaxy_ng.tests.integration.constants import USERNAME_PUBLISHER
+from ..utils.iqe_utils import require_signature_for_approval
 
 from ..utils import (
     CapturingGalaxyError,
@@ -341,6 +342,9 @@ def test_long_field_values(ansible_config, upload_artifact, field):
 @pytest.mark.importer
 @pytest.mark.min_hub_version("4.6dev")
 @pytest.mark.all
+@pytest.mark.skipif(require_signature_for_approval(), reason="This test needs refactoring to "
+                                                             "work with signatures required "
+                                                             "on move.")
 def test_ansible_requires(ansible_config, upload_artifact, spec, settings):
     """
     Test handling of POSTs to the artifact endpoint neglecting to submit a file.
@@ -348,8 +352,6 @@ def test_ansible_requires(ansible_config, upload_artifact, spec, settings):
     Also verifies that the collections endpoint properly returns a `requires_ansible` field,
     and that the returned field matches the collection metadata.
     """
-    if settings.get("GALAXY_REQUIRE_SIGNATURE_FOR_APPROVAL"):
-        pytest.skip("This test needs refactoring to work with signatures required on move.")
 
     config = ansible_config("basic_user")
     api_client = get_client(config)
