+++ galaxy_ng/tests/integration/utils/iqe_utils.py
@@ -1,14 +1,17 @@
 """Utility functions for AH tests."""
 import os
 import subprocess
+from functools import lru_cache
 from unittest.mock import patch
 
-from galaxy_ng.tests.integration.constants import BETA_GALAXY_STAGE_PROFILES
+from galaxy_ng.tests.integration.constants import BETA_GALAXY_STAGE_PROFILES, SYNC_PROFILES, \
+    EPHEMERAL_PROFILES, DEPLOYED_PAH_PROFILES, PROFILES, CREDENTIALS
+from galaxy_ng.tests.integration.utils import get_client
 
 from galaxykit import GalaxyClient
 
 import logging
-from urllib.parse import urljoin
+from urllib.parse import urljoin, urlparse
 
 from ansible.galaxy.api import GalaxyAPI
 from ansible.galaxy.token import BasicAuthToken
@@ -335,3 +338,297 @@ def beta_galaxy_user_cleanup(gc, u):
         delete_v1_namespace(gc_admin, github_user_username)
     except ValueError:
         pass
+
+
+def get_ansible_config():
+    return AnsibleConfigFixture
+
+
+def get_galaxy_client(ansible_config):
+    """
+    Returns a function that, when called with one of the users listed in the settings.local.yaml
+    file will login using hub and galaxykit, returning the constructed GalaxyClient object.
+    """
+    galaxy_kit_client = GalaxyKitClient(ansible_config)
+    return galaxy_kit_client.gen_authorized_client
+
+
+class AnsibleConfigFixture(dict):
+    # The class is instantiated with a "profile" that sets
+    # which type of user will be used in the test
+
+    PROFILES = {}
+
+    def __init__(self, profile=None, namespace=None, url=None, auth_url=None):
+        backend_map = {
+            "community": "community",
+            "galaxy": "galaxy",
+            "keycloak": "ldap",
+            "ldap": "ldap"
+        }
+        self._auth_backend = os.environ.get('HUB_TEST_AUTHENTICATION_BACKEND')
+        self.url = url
+        self.auth_url = auth_url
+        self.profile = profile
+        self.namespace = namespace
+
+        if is_sync_testing():
+            self.PROFILES = SYNC_PROFILES
+        elif is_stage_environment():
+            self.PROFILES = EPHEMERAL_PROFILES
+        elif not is_dev_env_standalone():
+            self.PROFILES = DEPLOYED_PAH_PROFILES
+            self._set_credentials_when_not_docker_pah()
+        elif is_ephemeral_env():
+            self.PROFILES = DEPLOYED_PAH_PROFILES
+            self.PROFILES["admin"]["token"] = None
+            self.PROFILES["org_admin"]["token"] = None
+            self.PROFILES["partner_engineer"]["token"] = None
+            self.PROFILES["basic_user"]["token"] = None
+        elif is_beta_galaxy_stage():
+            self.PROFILES = BETA_GALAXY_STAGE_PROFILES
+        else:
+            for profile_name in PROFILES:
+                p = PROFILES[profile_name]
+                credential_set = backend_map.get(self._auth_backend, "galaxy")
+                if p['username'] is None:
+                    continue
+
+                if username := p["username"].get(credential_set):
+                    self.PROFILES[profile_name] = {
+                        "username": username,
+                        "token": CREDENTIALS[username].get("token"),
+                        "password": CREDENTIALS[username].get("password")
+                    }
+
+            if self._auth_backend == "community":
+                self.PROFILES["anonymous_user"] = PROFILES.get('anonymous_user')
+
+        # workaround for a weird error with the galaxy cli lib ...
+        galaxy_token_fn = os.path.expanduser('~/.ansible/galaxy_token')
+        if not os.path.exists(os.path.dirname(galaxy_token_fn)):
+            os.makedirs(os.path.dirname(galaxy_token_fn))
+        if not os.path.exists(galaxy_token_fn):
+            with open(galaxy_token_fn, 'w') as f:
+                f.write('')
+
+        if profile:
+            self.set_profile(profile)
+
+    def __hash__(self):
+        # To avoid TypeError: unhashable type: 'AnsibleConfigFixture'
+        return hash((self.url, self.auth_url, self.profile, self.namespace))
+
+    def _set_profile_from_vault(self, loader, profile, param):
+        param_vault_path = self.PROFILES[profile][param]["vault_path"]
+        param_vault_key = self.PROFILES[profile][param]["vault_key"]
+        param_from_vault = loader.get_value_from_vault(path=param_vault_path,
+                                                       key=param_vault_key)
+        self.PROFILES[profile][param] = param_from_vault
+
+    def _set_credentials_when_not_docker_pah(self):
+        # if we get here, we are running tests against PAH
+        # but not in a containerized development environment (probably from AAP installation),
+        # so we need to get the URL and admin credentials (and create some test data)
+        admin_pass = os.getenv("HUB_ADMIN_PASS", "AdminPassword")
+        self.PROFILES["admin"]["username"] = "admin"
+        self.PROFILES["admin"]["password"] = admin_pass
+        self.PROFILES["admin"]["token"] = None
+        self.PROFILES["iqe_admin"]["username"] = "admin"
+        self.PROFILES["iqe_admin"]["password"] = admin_pass
+        self.PROFILES["iqe_admin"]["token"] = None
+        self.PROFILES["basic_user"]["token"] = None
+        self.PROFILES["basic_user"]["password"] = "Th1sP4ssd"
+        self.PROFILES["partner_engineer"]["token"] = None
+        self.PROFILES["partner_engineer"]["password"] = "Th1sP4ssd"
+        self.PROFILES["ee_admin"]["token"] = None
+        self.PROFILES["ee_admin"]["password"] = "Th1sP4ssd"
+        self.PROFILES["org_admin"]["token"] = None
+        self.PROFILES["org_admin"]["password"] = "Th1sP4ssd"
+        token = get_standalone_token(self.PROFILES["admin"], server=self.get("url"),
+                                     ssl_verify=False)
+        self.PROFILES["admin"]["token"] = token
+
+    def __repr__(self):
+        return f'<AnsibleConfigFixture: {self.namespace}>'
+
+    def __getitem__(self, key):
+
+        if key == 'url':
+            # The "url" key is actually the full url to the api root.
+            if self.url:
+                return self.url
+            else:
+                return os.environ.get(
+                    'HUB_API_ROOT',
+                    'http://localhost:5001/api/automation-hub/'
+                )
+        elif key == 'api_prefix':
+            # strip the proto+host+port from the api root
+            api_root = os.environ.get(
+                'HUB_API_ROOT',
+                'http://localhost:5001/api/automation-hub/'
+            )
+            parsed = urlparse(api_root)
+            return parsed.path
+
+        elif key == 'auth_url':
+            # The auth_url value should be None for a standalone stack.
+            if self.auth_url:
+                return self.auth_url
+            else:
+                return os.environ.get(
+                    'HUB_AUTH_URL',
+                    None
+                )
+
+        elif key == 'auth_backend':
+            return self._auth_backend
+
+        elif key == "token":
+            # Generate tokens for LDAP and keycloak backed users
+            if self.profile:
+                p = self.PROFILES[self.profile]
+                try:
+                    if CREDENTIALS[p["username"]].get("gen_token", False):
+                        return get_standalone_token(p, self["url"])
+                    return p.get("token", None)
+                except KeyError:
+                    return p.get("token", None)
+            else:
+                return None
+
+        elif key == "username":
+            return self.PROFILES[self.profile]["username"]
+
+        elif key == "password":
+            return self.PROFILES[self.profile]["password"]
+
+        elif key == 'use_move_endpoint':
+            # tells the tests whether or not to try to mark
+            # an imported collection as "published". This happens
+            # automatically in the default config for standalone,
+            # so should return False in that case ...
+
+            if os.environ.get('HUB_USE_MOVE_ENDPOINT'):
+                val = os.environ['HUB_USE_MOVE_ENDPOINT']
+                if str(val) in ['1', 'True', 'true']:
+                    return True
+
+            # standalone ...
+            return False
+
+            # cloud ...
+            # return True
+
+        elif key == 'upload_signatures':
+            if os.environ.get('HUB_UPLOAD_SIGNATURES'):
+                val = os.environ['HUB_UPLOAD_SIGNATURES']
+                if str(val) in ['1', 'True', 'true']:
+                    return True
+            return False
+
+        elif key == 'github_url':
+            return os.environ.get(
+                'SOCIAL_AUTH_GITHUB_BASE_URL',
+                'http://localhost:8082'
+            )
+
+        elif key == 'github_api_url':
+            return os.environ.get(
+                'SOCIAL_AUTH_GITHUB_API_URL',
+                'http://localhost:8082'
+            )
+
+        elif key == 'ssl_verify':
+            return os.environ.get(
+                'SSL_VERIFY',
+                False
+            )
+
+        elif key == 'container_engine':
+            return os.environ.get(
+                'CONTAINER_ENGINE',
+                'podman'
+            )
+
+        elif key == 'container_registry':
+            return os.environ.get(
+                'CONTAINER_REGISTRY',
+                'localhost:5001'
+            )
+
+        elif key == 'server':
+            return self["url"].split("/api/")[0]
+
+        elif key == 'remote_hub':
+            # The "url" key is actually the full url to the api root.
+            return os.environ.get(
+                'REMOTE_HUB',
+                'https://console.stage.redhat.com/api/automation-hub/'
+            )
+        elif key == 'remote_auth_url':
+            # The "url" key is actually the full url to the api root.
+            return os.environ.get(
+                'REMOTE_AUTH_URL',
+                'https://sso.stage.redhat.com/auth/realms/'
+                'redhat-external/protocol/openid-connect/token/'
+            )
+        elif key == 'local_hub':
+            # The "url" key is actually the full url to the api root.
+            return os.environ.get(
+                'LOCAL_HUB',
+                'http://localhost:5001/api/automation-hub/'
+            )
+        elif key == 'local_auth_url':
+            # The "url" key is actually the full url to the api root.
+            return os.environ.get(
+                'LOCAL_AUTH_URL',
+                None
+            )
+
+        else:
+            raise Exception(f'Unknown config key: {self.namespace}.{key}')
+
+        return super().__getitem__(key)
+
+    def get(self, key):
+        return self.__getitem__(key)
+
+    def get_profile_data(self):
+        if self.profile:
+            return self.PROFILES[self.profile]
+        raise Exception("No profile has been set")
+
+    def set_profile(self, profile):
+        self.profile = profile
+        if isinstance(self.PROFILES[self.profile]["username"], dict):
+            # credentials from vault
+            loader = get_vault_loader()
+            self._set_profile_from_vault(loader, self.profile, "username")
+            self._set_profile_from_vault(loader, self.profile, "password")
+            if self.PROFILES[self.profile]["token"]:
+                self._set_profile_from_vault(loader, self.profile, "token")
+
+
+@lru_cache()
+def get_vault_loader():
+    from .vault_loading import VaultSecretFetcher
+    vault_settings = {
+        'IQE_VAULT_VERIFY': True,
+        'IQE_VAULT_URL': 'https://vault.devshift.net',
+        'IQE_VAULT_GITHUB_TOKEN': os.environ.get('IQE_VAULT_GITHUB_TOKEN'),
+        'IQE_VAULT_ROLE_ID': os.environ.get('IQE_VAULT_ROLE_ID'),
+        'IQE_VAULT_SECRET_ID': os.environ.get('IQE_VAULT_SECRET_ID'),
+        'IQE_VAULT_LOADER_ENABLED': True,
+        'IQE_VAULT_MOUNT_POINT': 'insights'
+    }
+    return VaultSecretFetcher.from_settings(vault_settings)
+
+
+def require_signature_for_approval():
+    ansible_config = get_ansible_config()
+    config = ansible_config("admin")
+    api_client = get_client(config)
+    settings = api_client("_ui/v1/settings/")
+    return settings.get("GALAXY_REQUIRE_SIGNATURE_FOR_APPROVAL")
